Support opening a NuGet package from a package source
The Open Package dialog will now open the selected package or
packages into the IDE. The NuGet package will be downloaded in the
background and its metadata and files displayed. When downloading
a downloading label is displayed at the top of the package metadata
and a Cancel button is displayed which will stop the download.
The package metadata available when displaying the package in the
Open Package dialog is initially displayed for the NuGet package when
it is opened in the IDE. After the download completes this information
is updated.