Open dialog if NuGet package cannot be found
When a NuGet package is right clicked in the Solution window now the
global packages folder is first checked to see if that NuGet package
exists, if not then the solution's packages directory is checked.
If the NuGet package cannot be found then the Open Package dialog
is opened and a search is run for that NuGet package.