Handle missing properties when using reflection
Made the code that uses reflection to get NuGet package information
from the Solution window's nodes more defensive. It now ensures that
properties and fields are available before trying to get the values.
So if these change in the main NuGet and .NET Core addins the only
thing that will happen is the Open Package menu will be disabled.