Handle package version wildcards
An attempt is made to parse the wildcard using the NuGet VersionRange
and the minimum version is used when trying to open the NuGet package.
PackageReferenceNodes do not return the correct wildcard string.
Instead they return 10.0-- without the '*' wildcard parts. The
correct value could be obtained by changing the reflection code but
for now this is not done. What happens is that the NuGet package will
not be found since the wrong path is used for the version part, it
contains the '-' characters, and the NuGet dialog will be displayed.