Fix being unable to open packages dialog from metadata view
Clicking a hyperlink for a package in the metadata view would not
open the dialog due to a null reference in xwt when trying to
search the cursor to the end of the text in the search text box.

System.NullReferenceException: Object reference not set to an
instance of an object
  at Xwt.Mac.TextEntryBackend.set_CursorPosition (System.Int32 value)
  in xwt/Xwt.XamMac/Xwt.Mac/TextEntryBackend.cs:171
  at Xwt.TextEntry.set_CursorPosition (System.Int32 value)
  in xwt/Xwt/Xwt/TextEntry.cs:112
  at Xwt.TextEntry.set_CursorPosition(int)

Setting this after the dialog is shown fixes the problem.

Also set focus to the search text box after the dialog is displayed
otherwise the focus is not set.