ILRepack the NuGet assemblies
Avoid problems with NuGet assemblies being updated in VS Mac by
ILRepacking them into a NuGet.Merged.dll which is used by the addin.
This was done in the past but the ILRepack NuGet package being used
did not support .NET Core assemblies. Switching to the
ILRepack.MSBuild.Task allows the NuGet assemblies to be merged into
a single dll. For some reason the dll targets .NET 5 but this can
be used by the addin.

https://github.com/peters/ILRepack.MSBuild.Task