Add workaround for opening a nuget package breaking the IDE layout
Using File - Open NuGet Package... without first showing the IDE
window would result in the IDE window being opened without any layout
and the .nupkg file is not displayed. This seems to be a bug in the
Workbench.OpenDocument call when a controller is passed. The
DocumentManager calls Workbench.EnsureShown which does not call
Present, which would initialize the layout, since Window.Visible
is true even though the main window has not been displayed for the
first time.

Workaround this by calling Workbench.Present if no layouts have been
created yet. This ensures the layouts are initialized before the
.nupkg file view is opened in the IDE.